This commit was generated by cvs2svn to compensate for changes in r11255, which included commits to RCS files with non-trunk default branches.
git-svn-id: http://svn.cern.ch/guest/lhcb/Hlt/trunk@77712 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 57174d382adea50dc511f16d224850417322d094

+++ Hlt/Hlt/HltDisplVertices/src/Hlt2DisplVerticesDEV.cpp
@@ -33,7 +33,7 @@ using namespace std;
 //-----------------------------------------------------------------------------
 // Implementation file for class : Hlt2DisplVerticesDEV
 //
-// 2009-july-1 : Neal Gueissaz
+// 2009-july-1 : Neal Gauvin (Gueissaz)
 //-----------------------------------------------------------------------------
 
 DECLARE_ALGORITHM_FACTORY(Hlt2DisplVerticesDEV);
@@ -241,7 +241,7 @@ StatusCode Hlt2DisplVerticesDEV::finalize() {
   if (msgLevel(MSG::DEBUG)) {
     debug() << "==> Finalize" << endmsg;
     if(m_nbevent == 0) m_nbevent++;
-    double err = 10.*sqrt( m_nbpassed/m_nbevent );
+    double err = 10.*sqrt( double(m_nbpassed/m_nbevent) );
     debug() << "------------- Efficiency -----------"<< endmsg;
     debug() << "| Accepted event         "<< 100.*m_nbpassed/m_nbevent 
 	    <<"+-" << err
